+++ CHANGELOG.md
@@ -1,6 +1,11 @@
 # Change Log
 
 ## [Unreleased]
+- Rollback relationships for model, use function:
+  - `changedBelongsTo` - to find out what `belongsTo` relationships changed
+  - `rollbackBelongsTo` - to perform rollback `belongsTo` relationships
+  - `changedHasMany` - to find out what `hasMany` relationships changed
+  - `rollbackHasMany` - to perform rollback `hasMany` relationships
 
 ## [0.5.0-beta.2] - 2016-08-23
 ### Changed